+++ optalg/opt_solver/__init__.py
@@ -1,7 +1,7 @@
 #****************************************************#
 # This file is part of OPTALG.                       #
 #                                                    #
-# Copyright (c) 2015-2017, Tomas Tinoco De Rubira.   #
+# Copyright (c) 2019, Tomas Tinoco De Rubira.        #
 #                                                    #
 # OPTALG is released under the BSD 2-clause license. #
 #****************************************************#
@@ -12,11 +12,13 @@ from .problem_lin import LinProblem
 from .problem_mixintlin import MixIntLinProblem
 
 from .clp import OptSolverClp
+from .clp_cmd import OptSolverClpCMD
 from .cbc import OptSolverCbc
+from .cbc_cmd import OptSolverCbcCMD
 from .iqp import OptSolverIQP
 from .inlp import OptSolverINLP
 from .ipopt import OptSolverIpopt
 from .augl import OptSolverAugL
 from .nr import OptSolverNR
-from .opt_solver_error import OptSolverError
+from .opt_solver_error import *
 from .opt_solver import OptSolver, OptCallback, OptTermination