+++ optalg/opt_solver/problem_mixintlin.py
@@ -75,3 +75,59 @@ class MixIntLinProblem(OptProblem):
         print('A shape : (%d,%d)' %(self.A.shape[0],self.A.shape[1]))
         print('A nnz   : %.2f %%' %(100.*self.A.nnz/(self.A.shape[0]*self.A.shape[1])))
         print('integer : %d' %(np.sum(self.P)))
+
+    def write_to_lp_file(self, filename):
+
+        f = open(filename, 'w')
+
+        # Objective
+        f.write('Minimize\n')
+        row = ' obj: '
+        first = True
+        for i in np.where(self.c != 0.)[0]:
+            ci = self.c[i]
+            if first and ci > 0:
+                pre = ''
+            elif ci > 0:
+                pre = '+ '
+            else:
+                pre = '- '
+            row += '%s%.10e x%d ' %(pre, np.abs(ci), i)
+            first = False
+        f.write(row+'\n')
+
+        # Constraints
+        f.write('Subject to\n')
+        A = self.A.tocsr()
+        for i in range(A.shape[0]):
+            first = True
+            row = ' c%d: ' %i
+            for k in range(A.indptr[i], A.indptr[i+1]):
+                j = A.indices[k]
+                d = A.data[k]
+                b = self.b[i]
+                if first and d > 0:
+                    pre = ''
+                elif d > 0:
+                    pre = '+ '
+                else:
+                    pre = '- '
+                row += '%s%.10e x%d ' %(pre, np.abs(d), j)
+                first = False
+            row += '= %.10e' %b
+            f.write(row+'\n')
+
+        # Bounds
+        f.write('Bounds\n')
+        for i in range(self.c.size):
+            f.write(' %.10e <= x%d <= %.10e\n' %(self.l[i], i, self.u[i]))
+
+        # General
+        f.write('General\n')
+        row = ' '
+        for i in np.where(self.P)[0]:
+            row += 'x%d ' %i
+        f.write(row+'\n')
+        f.write('End\n')
+        
+        f.close()