+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Unreleased
 * Simplified barrier parameters in augl.
 * Added problem properties and methods for checking whether solvers support certain problem properties.
 * Added methods to convert general problem into linear, mixed integer linear, and quadratic problems.
+* Added methods to write lp files from linear and mixed integer linear problems.
+* Added command-line interfaces for cbc and clp.
 
 Version 1.1.7
 -------------